✨ 캐릭터 뽑기 모달 구현
- 첫 번째 모달 : 간단한 공지
- 두 번째 모달 : 뽑기 UI (룰렛 느낌)
